refactor(interface-layer): rename interactions package and add ChannelApi interface
