+++ servers-service/src/main/kotlin/piperkt/services/servers/interfaces/web/api/ServerHttpControllerApi.kt
@@ -5,7 +5,7 @@ import io.micronaut.http.annotation.PathVariable
 import io.micronaut.security.annotation.Secured
 import io.micronaut.security.rules.SecurityRule
 import java.security.Principal
-import piperkt.services.servers.interfaces.web.api.servers.ServerApi
+import piperkt.services.servers.interfaces.web.api.interactions.ServerApi
 
 @Secured(SecurityRule.IS_AUTHENTICATED)
 interface ServerHttpControllerApi {
