+++ servers-service/src/test/kotlin/piperkt/services/servers/interfaces/web/ServerControllerClient.kt
@@ -12,7 +12,7 @@ import io.micronaut.http.annotation.Post
 import io.micronaut.http.annotation.Put
 import io.micronaut.http.client.annotation.Client
 import io.micronaut.retry.annotation.Retryable
-import piperkt.services.servers.interfaces.web.api.servers.ServerApi
+import piperkt.services.servers.interfaces.web.api.interactions.ServerApi
 
 @Client("/servers")
 @Retryable
